players Move NOWW

+++ lib/Player/object_ball.py
@@ -10,8 +10,8 @@ class BallObject(Object):
 
     def init_str(self, string: str):
         data = string.split(" ")
-        self._pos = Vector2D(int(data[0]), int(data[1]))
-        self._vel = Vector2D(int(data[2]), int(data[3]))
+        self._pos = Vector2D(float(data[0]), float(data[1]))
+        self._vel = Vector2D(float(data[2]), float(data[3]))
 
     def __repr__(self):
         return f"(pos: {self.pos()}) (vel:{self.vel()})"
